sb3: fix asType and debug mode making everything break
If asType was given undefined, it would promptly decide to return the
entire script up to that point as the result. Unsurprisingly, this
caused a lot of issues. Only important change was line 5972, others are
to maybe make this more difficult.

If debug mode was enabled, CompiledExpression objects would wrongly be
stringified to '/* opcode_here */[object Object]', which is, of course,
problematic.