implement say() on Stage
Fixes crashes that come from trying to use say on a stage
Also fixes control flow that might for some reason use say for delays.

say() on a Scratch2Stage: no effects
say() on a Scratch3Stage: speech bubble at (0, 0)

Also fixed the color of text in the bubbles to be black instead of gray.