Remove timer caching
This commit reverts 8a860f08be537131e799342fa18663cf94ac4dc2

This makes preciseTimers the default and only option.
This fixes many projects that would otherwise get stuck in endless loops.
This includes all the gb.sb2 projects.

The original commit was pushed to nathan/phosphorus,
but it was never pushed to phosphorus/phosphorus.github.io,
so it never was implemented on the website.
There was a good reason for that, apparently.