sb3: Improve handling of NaN to match Scratch (#263)
NaN will now be converted to 0 when it is used, not when it is set.
Previously, anything that could be NaN would be eagerly converted to 0. This broke things that expected NaN.
The flag system might allow more flexibility in the future for things like nullability.

Also added a test.