sb3: Support event_whengreaterthan
Supports timer and loudness.

Should work basically the same as Scratch, although I haven't looked at what scratch-vm does.
I would probably become very upset with how simple this is to do in scratch-vm compared to what we have to deal with.

Anyways, it uses weird script preprocessing to conditionally execute the script.
The script is executed if the sensor condition is true, otherwise it just waits for the next tick.

If the function is executed, it enters a state where it will wait for the condition to be false before executing the script again when the condition is true.
This is based on observations of what this script does in practice.