Ensure that P.io.Img can get gc'd
It's been found that the Img class can sometimes refuse to be GC'd in some browsers
Sometimes we have thousands of these, and that adds up
Resetting the image event handlers to null appears to resolve this
To be clear, this is about our `Img` class, not the actual images. Those can't get GC'd because we need them for rendering.