sb2: always remove SVG from document
Previously, a malicious SVG could theoretically induce
patchSVG into throwing an error which would cause the SVG
to remain in the document indefinitely. It goes through
DOMPurify already so this is safe, but as an extra precaution,
we now always remove the SVG from the document, even in
the event of an error.

This does not fix any known security bugs. It is just a
precaution.